moved microsoft graph service out of uwp services project
