+++ Microsoft.Toolkit.Services/Services/MicrosoftGraph/MicrosoftGraphAuthenticationHelper.cs
@@ -12,8 +12,15 @@
 
 using System;
 using System.Linq;
+#if WINRT
+using System.Net.Http;
+#endif
 using System.Threading.Tasks;
 using Microsoft.Identity.Client;
+#if WINRT
+using Microsoft.IdentityModel.Clients.ActiveDirectory;
+using Windows.Storage;
+#endif
 using MSAL = Microsoft.Identity.Client;
 
 namespace Microsoft.Toolkit.Services.MicrosoftGraph
@@ -33,7 +40,30 @@ namespace Microsoft.Toolkit.Services.MicrosoftGraph
         protected const string AuthorizationTokenService = "https://login.microsoftonline.com/common/oauth2/v2.0/token";
         protected const string LogoutUrlV2Model = "https://login.microsoftonline.com/common/oauth2/v2.0/logout";
 
+#if WINRT
+        private const string LogoutUrl = "https://login.microsoftonline.com/common/oauth2/logout";
+        private const string MicrosoftGraphResource = "https://graph.microsoft.com";
+
+        /// <summary>
+        /// Storage key name for user name.
+        /// </summary>
+        private static readonly string STORAGEKEYUSER = "user";
+
+#endif
+
         private static MSAL.PublicClientApplication _identityClient = null;
+
+#if WINRT
+        /// <summary>
+        /// Password vault used to store access tokens
+        /// </summary>
+        private readonly Windows.Security.Credentials.PasswordVault _vault;
+
+        /// <summary>
+        /// Azure Active Directory Authentication context use to get an access token [ADAL]
+        /// </summary>
+        private AuthenticationContext _azureAdContext = new AuthenticationContext(Authority);
+#endif
 
         /// <summary>
         /// Gets or sets delegated permission Scopes
@@ -45,13 +75,20 @@ namespace Microsoft.Toolkit.Services.MicrosoftGraph
         /// </summary>
         public MicrosoftGraphAuthenticationHelper()
         {
+#if WINRT
+            _vault = new Windows.Security.Credentials.PasswordVault();
+#endif
         }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="MicrosoftGraphAuthenticationHelper"/> class.
         /// </summary>
         /// <param name="delegatedPermissionScopes">Delegated Permission Scopes</param>
-        public MicrosoftGraphAuthenticationHelper(string[] delegatedPermissionScopes) => DelegatedPermissionScopes = delegatedPermissionScopes;
+        public MicrosoftGraphAuthenticationHelper(string[] delegatedPermissionScopes)
+            : this()
+        {
+            DelegatedPermissionScopes = delegatedPermissionScopes;
+        }
 
         /// <summary>
         /// Gets or sets the Oauth2 access token.
@@ -66,10 +103,14 @@ namespace Microsoft.Toolkit.Services.MicrosoftGraph
         /// <summary>
         /// Clean the TokenCache
         /// </summary>
-        internal virtual void CleanToken()
+        internal void CleanToken()
         {
             TokenForUser = null;
+#if WINRT
+            _azureAdContext.TokenCache.Clear();
+#endif
         }
+
         /// <summary>
         /// Get a Microsoft Graph access token using the v2.0 Endpoint
         /// </summary>
@@ -138,5 +179,61 @@ namespace Microsoft.Toolkit.Services.MicrosoftGraph
 
             return true;
         }
+
+#if WINRT
+        /// <summary>
+        /// Get a Microsoft Graph access token from Azure AD.
+        /// </summary>
+        /// <param name="appClientId">Azure AD application client ID</param>
+        /// <returns>An oauth2 access token.</returns>
+        internal async Task<string> GetUserTokenAsync(string appClientId)
+        {
+            // For the first use get an access token prompting the user, after one hour
+            // refresh silently the token
+            if (TokenForUser == null)
+            {
+                IdentityModel.Clients.ActiveDirectory.AuthenticationResult userAuthnResult = await _azureAdContext.AcquireTokenAsync(MicrosoftGraphResource, appClientId, new Uri(DefaultRedirectUri), new IdentityModel.Clients.ActiveDirectory.PlatformParameters(PromptBehavior.Always, false));
+                TokenForUser = userAuthnResult.AccessToken;
+                Expiration = userAuthnResult.ExpiresOn;
+            }
+
+            if (Expiration <= DateTimeOffset.UtcNow.AddMinutes(5))
+            {
+                IdentityModel.Clients.ActiveDirectory.AuthenticationResult userAuthnResult = await _azureAdContext.AcquireTokenSilentAsync(MicrosoftGraphResource, appClientId);
+                TokenForUser = userAuthnResult.AccessToken;
+                Expiration = userAuthnResult.ExpiresOn;
+            }
+
+            return TokenForUser;
+        }
+
+        /// <summary>
+        /// Logout the user
+        /// </summary>
+        /// <param name="authenticationModel">Authentication version endPoint</param>
+        /// <returns>Success or failure</returns>
+        internal async Task<bool> LogoutAsync(string authenticationModel)
+        {
+            HttpResponseMessage response = null;
+            ApplicationData.Current.LocalSettings.Values[STORAGEKEYUSER] = null;
+
+            if (authenticationModel.Equals("V1"))
+            {
+                using (var client = new HttpClient())
+                {
+                    var request = new HttpRequestMessage(HttpMethod.Get, LogoutUrl);
+                    response = await client.SendAsync(request);
+                    return response.IsSuccessStatusCode;
+                }
+            }
+
+            if (authenticationModel.Equals("V2"))
+            {
+                return Logout();
+            }
+
+            return true;
+        }
+#endif
     }
 }
