+++ Microsoft.Toolkit.Services/Services/MicrosoftGraph/User/MicrosoftGraphUserService.cs
@@ -36,6 +36,17 @@ namespace Microsoft.Toolkit.Services.MicrosoftGraph
             PhotosService = photosService;
         }
 
+#if WINRT
+        /// <summary>
+        /// Initializes a new instance of the <see cref="MicrosoftGraphUserService"/> class.
+        /// </summary>
+        /// <param name="graphProvider">Instance of GraphClientService class</param>
+        public MicrosoftGraphUserService(GraphServiceClient graphProvider)
+            : this(graphProvider, new MicrosoftGraphUserServicePhotos(graphProvider))
+        {
+        }
+#endif
+
         ///// <summary>
         ///// MicrosoftGraphServiceMessages instance
         ///// </summary>
@@ -101,6 +112,27 @@ namespace Microsoft.Toolkit.Services.MicrosoftGraph
 
             return _currentConnectedUser;
         }
+
+#if WINRT
+        /// <summary>
+        /// Retrieve current connected user's photo.
+        /// </summary>
+        /// <param name="cancellationToken">The <see cref="CancellationToken"/> for the request.</param>
+        /// <returns>A stream containing the user"s photo</returns>
+        public async Task<Windows.Storage.Streams.IRandomAccessStream> GetPhotoAsync(CancellationToken cancellationToken)
+        {
+            return (await PhotosService.GetPhotoAsync(CancellationToken.None)) as Windows.Storage.Streams.IRandomAccessStream;
+        }
+
+        /// <summary>
+        /// Retrieve current connected user's photo.
+        /// </summary>
+        /// <returns>A stream containing the user"s photo</returns>
+        public async Task<Windows.Storage.Streams.IRandomAccessStream> GetPhotoAsync()
+        {
+            return (await PhotosService.GetPhotoAsync()) as Windows.Storage.Streams.IRandomAccessStream;
+        }
+#endif
 
         /// <summary>
         /// Create an instance of MicrosoftGraphServiceMessage
