--- Microsoft.Toolkit.Uwp.Services/Services/MicrosoftGraph/Events/MicrosoftGraphServiceEvent.cs
@@ -1,35 +1,0 @@
-﻿// ******************************************************************
-// Copyright (c) Microsoft. All rights reserved.
-// This code is licensed under the MIT License (MIT).
-// THE CODE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
-// INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
-// IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
-// DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
-// TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH
-// THE CODE OR THE USE OR OTHER DEALINGS IN THE CODE.
-// ******************************************************************
-
-using System;
-using Microsoft.Graph;
-
-namespace Microsoft.Toolkit.Uwp.Services.MicrosoftGraph
-{
-    /// <summary>
-    ///  Class using Office 365 Microsoft Graph Messages API
-    /// </summary>
-    [Obsolete("This class is being deprecated. Please use the .NET Standard Library counterpart found in Microsoft.Toolkit.Services.")]
-
-    public class MicrosoftGraphServiceEvent : Toolkit.Services.MicrosoftGraph.MicrosoftGraphServiceEvent
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="MicrosoftGraphServiceEvent"/> class.
-        /// </summary>
-        /// <param name="graphClientProvider">Instance of GraphClientService class</param>
-        /// <param name="currentConnectedUser">Instance of Graph.User class</param>
-        public MicrosoftGraphServiceEvent(GraphServiceClient graphClientProvider, User currentConnectedUser)
-            : base(graphClientProvider, currentConnectedUser)
-        {
-        }
-    }
-}
